+++ Sources/TuistPluginSwiftLintFramework/Constants/Constants.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-/// A set of constant values that can be reused across `TuistPluginSwiftLintFramework` module.
-enum Constants {
-    /// The plugin version.
-    static let version = "0.2.0"
-}