+++ CHANGELOG.md
@@ -8,8 +8,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ### Changed
  
-- `SwiftLint` dependency has been updated to `0.46.5` (was `0.46.1`).
-- `tuist` dependency has been replaced with `ProjectAutomation`.
+- support `tuist` scaffold
 
 ## 0.1.0